Pull series show text into module
