+++ app/modules/series_show_text.rb
@@ -1,0 +1,15 @@
+module SeriesShowText
+
+  def title
+    "SERIES PROFILE"
+  end
+
+  def relation_1
+    "SERIES CHARACTERS"
+  end
+
+  def relation_2
+    "SERIES EVENTS"
+  end
+  
+end
